requirements.txt
google-adk>=0.3.0
google-genai>=0.8.0
google-cloud-firestore>=2.16.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
